+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 - Add Chisel Reborn
 - Add Factory Blocks
 - Add Luminax
+- Add GregTech CEu: Modern
+- Add mining dimensions
 - Remove Iris/Oculus & GeckoLib Compat
 - Remove Structurize
+- Remove The Afterdark
+- Do not mark mods as server-only, because packwiz doesn't export them to client-pack
 - Mark Model Gap Fix as client-only
 - Mark Connected Glass as client-only
 - Mark Dark Mode Everywhere as client-only
@@ -13,6 +17,7 @@
 - Mark ImmediatelyFast as client-only
 - Mark Presence Footsteps as client-only
 - Mark Mouse Tweaks as client-only
+- Update mods
 
 ## 1.0.0-alpha.12